requirements.txt
a2wsgi==1.10.10
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiohttp-asgi-connector==1.1.2
aiosignal==1.4.0
aiosqlite==0.22.1
annotated-types==0.7.0
anyio==4.12.0
appdirs==1.4.4
asgiref==3.11.0
asttokens==3.0.0
attrs==25.4.0
beautifulsoup4==4.13.4
bentoml==1.4.33
black==25.11.0
blinker==1.9.0
cattrs==23.1.2
certifi==2025.7.14
cffi==2.0.0
charset-normalizer==3.4.2
click==8.3.1
click-option-group==0.5.9
cloudpickle==3.1.2
comm==0.2.3
cryptography==46.0.3
debugpy==1.8.17
decorator==5.2.1
deprecation==2.1.0
distro==1.9.0
dynaconf==3.2.12
evidently==0.7.20
executing==2.2.1
Faker==40.1.0
fastapi==0.115.12
filelock==3.20.3
Flask==3.1.2
frozenlist==1.8.0
fsspec==2026.1.0
greenlet==3.3.0
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
httpx-ws==0.8.2
idna==3.10
importlib_metadata==8.7.1
iniconfig==2.3.0
ipykernel==6.30.1
ipython==9.6.0
ipython_pygments_lexers==1.1.1
iterative-telemetry==0.0.10
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.1
jupyter_client==8.6.3
jupyter_core==5.8.1
kantoku==0.18.3
litestar==2.19.0
litestar-htmx==0.5.0
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib-inline==0.1.7
mdurl==0.1.2
msgspec==0.20.0
multidict==6.7.0
multipart==1.3.0
mypy_extensions==1.1.0
nest-asyncio==1.6.0
nltk==3.9.2
numpy==2.2.6
nvidia-ml-py==13.590.48
nvidia-nccl-cu12==2.30.4
opentelemetry-api==1.39.1
opentelemetry-instrumentation==0.60b1
opentelemetry-instrumentation-aiohttp-client==0.60b1
opentelemetry-instrumentation-asgi==0.60b1
opentelemetry-proto==1.39.1
opentelemetry-sdk==1.39.1
opentelemetry-semantic-conventions==0.60b1
opentelemetry-util-http==0.60b1
packaging==25.0
pandas==2.3.3
parso==0.8.5
pathspec==0.12.1
patsy==1.0.2
pexpect==4.9.0
pip-requirements-parser==32.0.1
platformdirs==4.5.0
playwright==1.57.0
plotly==5.24.1
pluggy==1.6.0
polyfactory==3.2.0
prometheus_client==0.24.1
prompt_toolkit==3.0.52
propcache==0.4.1
protobuf==6.33.4
psutil==7.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==22.0.0
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
pyee==13.0.0
Pygments==2.19.2
PyJWT==2.10.1
pyparsing==3.3.2
pytest==9.0.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-json-logger==4.0.0
python-multipart==0.0.21
pytokens==0.3.0
pytz==2025.2
PyYAML==6.0.3
pyzmq==27.1.0
regex==2025.11.3
requests==2.32.4
rich==14.2.0
rich-click==1.9.5
rich-toolkit==0.17.1
schema==0.7.8
scikit-learn==1.8.0
scipy==1.16.3
shellingham==1.5.4
simple-di==0.1.5
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
stack-data==0.6.3
starlette==0.46.2
statsmodels==0.14.6
tenacity==9.1.2
threadpoolctl==3.6.0
tomli_w==1.2.0
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
typer==0.21.1
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.14.1
tzdata==2025.3
ujson==5.11.0
urllib3==2.5.0
uuid6==2025.0.1
uvicorn==0.34.3
uvloop==0.22.1
watchdog==6.0.0
watchfiles==1.1.1
wcwidth==0.2.14
websockets==16.0
Werkzeug==3.1.4
wrapt==1.17.3
wsproto==1.3.2
xgboost==3.2.0
yarl==1.22.0
zipp==3.23.0
